deprecating and renaming files according to api versions
